el chupa ortos

+++ src/main/java/com/example/recetarium/demo/DTOs/IngredienteRequestDto.java
@@ -3,7 +3,7 @@ package com.example.recetarium.demo.DTOs;
 public class IngredienteRequestDto {
     private String nombreIngrediente; // debe existir en la BD
     private String unidad;            // debe existir en la BD
-    private int cantidad;
+    private float cantidad;
     private String observacion;
 
     public String getNombreIngrediente() {
@@ -14,7 +14,7 @@ public class IngredienteRequestDto {
         return unidad;
     }
 
-    public int getCantidad() {
+    public float getCantidad() {
         return cantidad;
     }
